emotion-en-service/requirements.txt
pykafka
avro-python3
requests
feedparser
osint-python-test-bed-adapter
osint-python-starter-service
textblob
transformers